python-client/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
Pillow==9.5.0
starlette==0.26.1
pydantic==1.10.7

